+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 
 - Document classes are now located under `app/documents`
 - Document layout are now located under `app/documents/layouts`
+- Invoice compileHeaderFooter will not compile bank and paypal anymore.